requirements.txt
python-docx
pandas
openpyxl
gspread
oauth2client
PyGithub
